Extend tests Wunderbyte-GmbH/Wunderbyte-GmbH#1648
